+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+### 2.19-beta5 / 2021-08-01
+
+- Bug fixes to autobind (#1035, David G. Young)
+
+### 2.19-beta4 / 2021-07-10
+
+- Auto start background power saver when using autobind.   Expedite region exits when using intent scan strategy.  (#1035, David G. Young)
+
+### 2.19-beta3 / 2021-07-04
+
+- Add experimental autobind methods for starting beacon ranging/monitoring without a manual call to bind(). (#1035, David G. Young)
+- Add support for beacon parsers based on 128-bit service UUIDs (#1035, David G. Young)
+
+
 ### 2.19-beta2 / 2021-06-07
 
 - Add experimental IntentScanStrategy to eliminate a foreground service for some background scan use cases. (#1030, David G. Young)